backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0
pymysql==1.1.0
python-dotenv==1.0.0
requests==2.31.0
cryptography==41.0.7
sqlparse==0.4.4
APScheduler==3.10.4
